requirements.txt
langchain
langchain-community
langchain-google-genai
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread
python-dotenv